move L2 readers to extras submodule
