+++ pyaerocom/extras/satellite_l2/base_reader.py
