+++ pyaerocom/extras/satellite_l2/sentinel5p.py
@@ -7,8 +7,9 @@ import time
 import numpy as np
 
 from pyaerocom import const
-from pyaerocom.io.readsatellitel2base import ReadL2DataBase
 from pyaerocom.ungriddeddata import UngriddedData
+
+from .base_reader import ReadL2DataBase
 
 
 class ReadL2Data(ReadL2DataBase):
